Fix 64-bit windows size_t assert
R=mstarzinger@chromium.org
TEST=Windows 64 bit build goes green

Review URL: https://chromiumcodereview.appspot.com/10221006

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@11438 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/hydrogen.cc
@@ -1630,7 +1630,7 @@ GVN_UNTRACKED_FLAG_LIST(DECLARE_FLAG)
 #else
   OS::SNPrintF(buffer, "0x%08X", flags.ToIntegral());
 #endif
-  uint32_t string_len = strlen(underlying_buffer) + 1;
+  size_t string_len = strlen(underlying_buffer) + 1;
   ASSERT(string_len <= sizeof(underlying_buffer));
   char* result = new char[strlen(underlying_buffer) + 1];
   memcpy(result, underlying_buffer, string_len);
